code update

+++ src/main/java/com/project/shop/member/service/MemberService.java
@@ -23,14 +23,33 @@ public class MemberService {
     private final AuthorityRepository authorityRepository;
     private final CartRepository cartRepository;
 
+    //회원 정보 전체 조회
+    public List<MemberResponse> userFindAll(String loginId){
 
+        Member member = findLoginMember(loginId);
+
+        Authority authority = authorityRepository.findByMember(member);
+
+        if(authority.getAuthName().equals("user"))
+            throw new RuntimeException("ONLY_ADMIN");
+
+        return memberRepository.findAll()
+                .stream()
+                .map(x -> {
+                    Authority auth = authorityRepository.findByMember(x);
+                    return MemberResponse.fromEntity(x,auth);
+                })
+                .toList();
+
+    }
 
     //내 정보 조회
     public MemberResponse userDetailFind(String loginId){
 
         Member member = findLoginMember(loginId);
 
-        return MemberResponse.fromEntity(member);
+        Authority authority = authorityRepository.findByMember(member);
+        return MemberResponse.fromEntity(member,authority);
     }
 
 
@@ -38,9 +57,12 @@ public class MemberService {
     public void update(String loginId, MemberUpdateRequest memberUpdateRequest){
 
         Member member = findLoginMember(loginId);
+        Member updateMember = member.updateUser(memberUpdateRequest);
+        memberRepository.save(updateMember);
 
-        Member update = member.updateUser(memberUpdateRequest);
-        memberRepository.save(update);
+        Authority authority = authorityRepository.findByMember(member);
+        Authority updateAuth = authority.updateAuth(memberUpdateRequest.auth());
+        authorityRepository.save(updateAuth);
     }
 
     //회원 탈퇴
@@ -48,16 +70,17 @@ public class MemberService {
 
         Member member = findLoginMember(loginId);
 
-        Member delete = member.deleteUser();
-        memberRepository.save(delete);
-
         //권한 삭제
         Authority authority = authorityRepository.findByMember(member);
         authorityRepository.delete(authority);
 
+        //member soft delete
+        memberRepository.delete(member);
+
         //장바구니 삭제
         List<Cart> cart = cartRepository.findAllByMember(member);
-        cartRepository.deleteAll(cart);
+        if (!cart.isEmpty())
+            cartRepository.deleteAll(cart);
 
     }
